Update version to 0.184.0 in deno.json and refactor cost function handling. Introduce a new CostInterface for better organization of cost functions, enhancing clarity and maintainability. Update imports across multiple files to reflect the new structure, ensuring consistent access to cost function definitions. Improve documentation for cost functions to provide clearer usage examples and characteristics.
