+++ deno.json
@@ -1,7 +1,7 @@
 {
   "name": "@stsoftware/neat-ai",
   "exports": "./mod.ts",
-  "version": "0.183.0",
+  "version": "0.184.0",
   "imports": {
     "@std/assert": "jsr:@std/assert@1.0.14",
     "@std/bytes": "jsr:@std/bytes@1.0.6",
