+++ src/Creature.ts
@@ -32,7 +32,8 @@ import { compactCreature } from "./compact/CompactCreature.ts";
 import { removeHiddenNeuron } from "./compact/CompactUtils.ts";
 import { createNeatConfig } from "./config/NeatConfig.ts";
 import type { NeatOptions } from "./config/NeatOptions.ts";
-import { type CostInterface, Costs } from "./Costs.ts";
+import { Costs } from "./Costs.ts";
+import type { CostInterface } from "./costs/CostInterface.ts";
 import { Activations } from "./methods/activations/Activations.ts";
 import { WorkerHandler } from "./multithreading/workers/WorkerHandler.ts";
 import { Neat } from "./NEAT/Neat.ts";
